Tc#3 create responsive carousel (#12)
* I have created own branch and after creating, I create own component for my task TS##

* TC-3
Fronted:
I installed npm react-slick for a beautiful carousel. I created a component for the future carousel and there is also a global state for it.

Backend:
I created a scheme for the carousel and there is also a router for CRUD requests, so far there are get and post requests in the router

* TC-3
Fronted:
I created a new component for admin and  there is also a new request, it is delete-request

* TC-3
Fronted:
I created a component in which the administration can edit a block with a slider, namely, delete and add photo frames. In the carousel block, I added a fixed height to the image

* TC-3
Fronted:
I added error checks so that the admin can
see them and there is also a limit on photos of at least 2.

updated the visual appearance of its components

Backend:
There is a put request for the admin with which you can replace images and there is
also fixtures for the carousel scheme

* I got a change from the new updated main branch and added them to my branch

* Tc-3
Fronted:
I corrected all the comments that the captain pointed out to me

* Tc-3
Fronted:
I used the prettier and also fixed the errors that the commander pointed out

* Tc-3
Fronted:
I fixed the errors that the commander pointed out

* Tc-3
Fronted:
I fixed the errors in Admin-panel for the slider that the commander pointed out

* TC#3-Create-Responsiv-Carousel

---------

Co-authored-by: Gordus Alexander <aldrmiller@gmail.com>